# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DSB-12 Python Developer Frameworks Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Framework {
  id: number;
  name: string;
  description: string;
  category: string;
  githubUrl?: string;
  pypiUrl?: string;
  officialUrl?: string;
  stars: number;
  weeklyDownloads: number;
  version: string;
  license: string;
  tags: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateFrameworkBody {
  name: string;
  description: string;
  category: string;
  githubUrl?: string;
  pypiUrl?: string;
  officialUrl?: string;
  version: string;
  license: string;
  tags: string[];
}

export interface Template {
  id: number;
  name: string;
  description: string;
  frameworkId: number;
  frameworkName: string;
  repoUrl?: string;
  language: string;
  difficulty: string;
  useCount: number;
  createdAt: string;
}

export interface CreateTemplateBody {
  name: string;
  description: string;
  frameworkId: number;
  repoUrl?: string;
  language: string;
  difficulty: string;
}

export interface StatsSummary {
  totalFrameworks: number;
  totalTemplates: number;
  totalStars: number;
  totalCategories: number;
  mostPopularCategory: string;
}

export interface CategoryStat {
  category: string;
  count: number;
  totalStars: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListFrameworksParams = {
  category?: string;
  search?: string;
};

export type ListTemplatesParams = {
  frameworkId?: number;
};