# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DSB-12 Python Developer Frameworks Hub API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all Python frameworks
 */
export const ListFrameworksQueryParams = zod.object({
  category: zod.coerce.string().optional(),
  search: zod.coerce.string().optional(),
});

export const ListFrameworksResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  githubUrl: zod.string().optional(),
  pypiUrl: zod.string().optional(),
  officialUrl: zod.string().optional(),
  stars: zod.number(),
  weeklyDownloads: zod.number(),
  version: zod.string(),
  license: zod.string(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const ListFrameworksResponse = zod.array(ListFrameworksResponseItem);

/**
 * @summary Add a new framework
 */
export const CreateFrameworkBody = zod.object({
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  githubUrl: zod.string().optional(),
  pypiUrl: zod.string().optional(),
  officialUrl: zod.string().optional(),
  version: zod.string(),
  license: zod.string(),
  tags: zod.array(zod.string()),
});

/**
 * @summary Get a framework by ID
 */
export const GetFrameworkParams = zod.object({
  id: zod.coerce.number(),
});

export const GetFrameworkResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  githubUrl: zod.string().optional(),
  pypiUrl: zod.string().optional(),
  officialUrl: zod.string().optional(),
  stars: zod.number(),
  weeklyDownloads: zod.number(),
  version: zod.string(),
  license: zod.string(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Update a framework
 */
export const UpdateFrameworkParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateFrameworkBody = zod.object({
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  githubUrl: zod.string().optional(),
  pypiUrl: zod.string().optional(),
  officialUrl: zod.string().optional(),
  version: zod.string(),
  license: zod.string(),
  tags: zod.array(zod.string()),
});

export const UpdateFrameworkResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  githubUrl: zod.string().optional(),
  pypiUrl: zod.string().optional(),
  officialUrl: zod.string().optional(),
  stars: zod.number(),
  weeklyDownloads: zod.number(),
  version: zod.string(),
  license: zod.string(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary Delete a framework
 */
export const DeleteFrameworkParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Star (upvote) a framework
 */
export const StarFrameworkParams = zod.object({
  id: zod.coerce.number(),
});

export const StarFrameworkResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  githubUrl: zod.string().optional(),
  pypiUrl: zod.string().optional(),
  officialUrl: zod.string().optional(),
  stars: zod.number(),
  weeklyDownloads: zod.number(),
  version: zod.string(),
  license: zod.string(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});

/**
 * @summary List all project templates
 */
export const ListTemplatesQueryParams = zod.object({
  frameworkId: zod.coerce.number().optional(),
});

export const ListTemplatesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  frameworkId: zod.number(),
  frameworkName: zod.string(),
  repoUrl: zod.string().optional(),
  language: zod.string(),
  difficulty: zod.string(),
  useCount: zod.number(),
  createdAt: zod.string(),
});
export const ListTemplatesResponse = zod.array(ListTemplatesResponseItem);

/**
 * @summary Create a new project template
 */
export const CreateTemplateBody = zod.object({
  name: zod.string(),
  description: zod.string(),
  frameworkId: zod.number(),
  repoUrl: zod.string().optional(),
  language: zod.string(),
  difficulty: zod.string(),
});

/**
 * @summary Get a template by ID
 */
export const GetTemplateParams = zod.object({
  id: zod.coerce.number(),
});

export const GetTemplateResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  frameworkId: zod.number(),
  frameworkName: zod.string(),
  repoUrl: zod.string().optional(),
  language: zod.string(),
  difficulty: zod.string(),
  useCount: zod.number(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a template
 */
export const DeleteTemplateParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get platform summary stats
 */
export const GetStatsSummaryResponse = zod.object({
  totalFrameworks: zod.number(),
  totalTemplates: zod.number(),
  totalStars: zod.number(),
  totalCategories: zod.number(),
  mostPopularCategory: zod.string(),
});

/**
 * @summary Framework count by category
 */
export const GetStatsByCategoryResponseItem = zod.object({
  category: zod.string(),
  count: zod.number(),
  totalStars: zod.number(),
});
export const GetStatsByCategoryResponse = zod.array(
  GetStatsByCategoryResponseItem,
);

/**
 * @summary Get top trending (most starred) frameworks
 */
export const GetTrendingFrameworksResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  description: zod.string(),
  category: zod.string(),
  githubUrl: zod.string().optional(),
  pypiUrl: zod.string().optional(),
  officialUrl: zod.string().optional(),
  stars: zod.number(),
  weeklyDownloads: zod.number(),
  version: zod.string(),
  license: zod.string(),
  tags: zod.array(zod.string()),
  createdAt: zod.string(),
  updatedAt: zod.string(),
});
export const GetTrendingFrameworksResponse = zod.array(
  GetTrendingFrameworksResponseItem,
);