# lib/api-zod/src/generated/types/createFrameworkBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DSB-12 Python Developer Frameworks Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateFrameworkBody {
  name: string;
  description: string;
  category: string;
  githubUrl?: string;
  pypiUrl?: string;
  officialUrl?: string;
  version: string;
  license: string;
  tags: string[];
}