# lib/api-zod/src/generated/types/createTemplateBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DSB-12 Python Developer Frameworks Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTemplateBody {
  name: string;
  description: string;
  frameworkId: number;
  repoUrl?: string;
  language: string;
  difficulty: string;
}