# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * DSB-12 Python Developer Frameworks Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Template {
  id: number;
  name: string;
  description: string;
  frameworkId: number;
  frameworkName: string;
  repoUrl?: string;
  language: string;
  difficulty: string;
  useCount: number;
  createdAt: string;
}